re-make branch + update migration

+++ app/Database/Migrations/2022-06-18-145837_JenisFaskes.php
@@ -10,7 +10,7 @@ class JenisFaskes extends Migration
     {
         $this->forge->addField([
             'id'          => [
-                'type'           => 'INT',
+                'type'           => 'BIGINT',
                 'constraint'     => 5,
                 'unsigned'       => true,
                 'auto_increment' => true
